fixed overflow error for large files

+++ lzwcpp/lzw.hh
@@ -8,7 +8,7 @@ class LZW {
     LZW() = default;
     ~LZW() = default;
 
-    void encode(const char* input_file, int file_size, std::ostream& output);
+    void encode(const char* input_file, uint64_t file_size, std::ostream& output);
     void decode(const char* input, std::ostream& output);
 
 };
